fix: rename + clean-up
